versioning, minor cleanup, save as defaults button
