+++ LogicReinc.BlendFarm/BlendFarmSettings.cs
@@ -1,10 +1,12 @@
-﻿using LogicReinc.BlendFarm.Server;
+﻿using LogicReinc.BlendFarm.Objects;
+using LogicReinc.BlendFarm.Server;
 using LogicReinc.BlendFarm.Shared;
 using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Text;
 using System.Text.Json;
+using static LogicReinc.BlendFarm.Windows.RenderWindow;
 
 namespace LogicReinc.BlendFarm
 {
@@ -23,6 +25,8 @@ namespace LogicReinc.BlendFarm
         /// Determines if the application should listen for servers over udp broadcasts
         /// </summary>
         public bool ListenForBroadcasts { get; set; } = true;
+
+        public OpenBlenderProject.UISettings UISettings { get; set; } = null;
 
 
         /// <summary>
